changed the templates folder structure and delete rendering in base.html. added the paths for links
