+++ config/routes.yaml
@@ -1,19 +1,17 @@
 index:
     path: /
-    controller: App\Controller\DefaultController::index
-
-app_lucky_number:
-  path: /lucky/number
-  controller: App\Controller\LuckyController::number
-
-donate:
-  path: /donate
-  controller: App\Controller\DonateController::donate
-
-author:
-  path: /author
-  controller: App\Controller\AuthorController::author
+    controller: App\Controller\ArticlesController::showAll
 
 article_show:
-  path: /article/{id}
-  controller: App\Controller\ArticleController::show
+    path: /article/{id}
+    controller: App\Controller\ArticlesController::show
+
+donate:
+    path: /donate
+    controller: App\Controller\DonateController::donate
+
+author:
+    path: /author
+    controller: App\Controller\AuthorController::author
+
+
