+++ templates/articles/author.html.twig
@@ -1,4 +1,4 @@
-{% extends 'pages/donate.html.twig' %}
+{% extends 'articles/donate.html.twig' %}
 
 {% block title %}About author{% endblock %}
